backend/requirements.txt
django==2.1.2
psycopg2-binary==2.7.5
coreapi==2.3.3
coreschema==0.0.4        
django-environ==0.4.4
django-auth-ldap==1.3.0
markdown==2.6.9
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
djangorestframework==3.7.3
django-cors-headers==2.1.0
django-filter==2.0.0
django-crispy-forms==1.7.2 
django_extensions==1.9.9
pylint-django==2.0
pylint-plugin-utils==0.4
pylint==2.0
openpyxl==2.5.0
ipython==6.2.1
gunicorn==19.7.1
pdbpp==0.9.2